+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * Made error messages from the lottie importer easier to understand
     * SVG export now supports hold frames
     * SVG import now assumes back for unspecified fill
+    * Ability to export as a spritesheet
 * Bug Fixes:
     * Fixed loading lottie animations with missing "a" attributes
     * The rotation handle no longer changes position based on scale